Create Winner_of_basketball_match.py
There are two basketball teams (Team1 and Team2) in a school and they play some matches everyday depending on their time and interest. Some days they play 3 matches, some days 2, some days 1 etc.

Write a python function, find_winner_of_the_day(), which accepts the name of the winner of each match and returns the name of the overall winner of the day. In case of equal number of wins, return “Tie”.